Refactor homepage. Extract balance history tabs. Move utils to module
