+++ src/views/Homepage/index.tsx
@@ -1,0 +1,27 @@
+import { Box, Flex } from '@chakra-ui/react'
+
+import Page from 'components/page/Page'
+import QuickTradeContainer from 'components/trade'
+
+import BalanceHistoryTabs from './BalanceHistoryTabs'
+
+const Homepage = () => {
+  return (
+    <Page>
+      <Flex
+        align={'center'}
+        direction={'column'}
+        justify={'center'}
+        mx='auto'
+        w={['340px', '500px', '820px', '1024px']}
+      >
+        <Box mb={12} w={['inherit', '500px']}>
+          <QuickTradeContainer />
+        </Box>
+        <BalanceHistoryTabs />
+      </Flex>
+    </Page>
+  )
+}
+
+export default Homepage
